Refactor scaling-mysql module into pipeline-mysql module
